Refactor move cleanup models into cleanup namespace
