+++ client/src/components/User/DiskUsage/Management/Cleanup/model/CleanupCategory.js
@@ -1,4 +1,4 @@
-import { CleanupOperation } from "../model";
+import { CleanupOperation } from ".";
 
 export class CleanupCategory {
     constructor(data) {
